% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{annotation_label_transfer}
\alias{annotation_label_transfer}
\title{Cell Type Annotation Transfer}
\usage{
annotation_label_transfer(
  input_read_RNA_assay,
  empty_droplets_tbl,
  reference_azimuth = NULL,
  assay = NULL
)
}
\arguments{
\item{input_read_RNA_assay}{SingleCellExperiment object containing RNA assay data.}

\item{empty_droplets_tbl}{A tibble identifying empty droplets.}

\item{reference_azimuth}{Optional reference data for Azimuth.}
<<<<<<< HEAD

\item{assay}{The assay to be used for analysis, specified as a character string.}
=======
>>>>>>> 7d4d87366996b9da235320e735f9434295a82dfd
}
\value{
A tibble with cell type annotation data.
}
\description{
\code{annotation_label_transfer} utilizes SingleR for cell-type identification using reference datasets
(Blueprint and Monaco Immune data). It can also perform cell type labeling using Azimuth when a reference
is provided.
}
