% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_UMAP}
\alias{calc_UMAP}
\title{Calculate UMAP Embeddings for Seurat Object}
\usage{
calc_UMAP(input_seurat)
}
\arguments{
\item{input_seurat}{Input data}
}
\value{
A tibble containing UMAP coordinates and cluster assignments, to be used for plotting and further analysis
}
\description{
Identify variable features, scale the data, run Principal Component Analysis (PCA), find neighbors, identify clusters,
and compute UMAP (Uniform Manifold Approximation and Projection) embeddings for
visualization of cell clusters.
}
