% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cell_communication}
\alias{cell_communication}
\title{Perform Human Cell-Cell Communication Analysis}
\usage{
cell_communication(
  input_read_RNA_assay,
  empty_droplets_tbl = NULL,
  alive_identification_tbl = NULL,
  doublet_identification_tbl = NULL,
  cell_type_tbl = NULL,
  assay = NULL,
  cell_type_column = NULL,
  feature_nomenclature,
  species_db,
  ...
)
}
\arguments{
\item{input_read_RNA_assay}{A SingleCellExperiment or Seurat object containing gene expression data}

\item{empty_droplets_tbl}{Optional tibble identifying empty droplets to be filtered out}

\item{alive_identification_tbl}{Optional tibble identifying dead cells to be filtered out}

\item{doublet_identification_tbl}{Optional A tibble from doublet identification.}

\item{cell_type_tbl}{Optional A tibble containing cell, cell type, and sample_id information}

\item{assay}{Character string specifying which assay to use}

\item{cell_type_column}{Character string specifying the column name containing cell type annotations}

\item{feature_nomenclature}{Character vector specifying gene in Symbol or Ensemble format}

\item{species_db}{The ligand-receptor interaction database curated in CellChat tool.}

\item{...}{Additional arguments passed to \code{CellChat::subsetDB}}
}
\value{
A CellChat tibble containing the inferred communication at the level of
ligands/receptors
}
\description{
This function performs cells communication analysis.
It processes single-cell RNA sequencing data to identify and analyze intercellular communication networks.
}
