% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cell_cycle_scoring}
\alias{cell_cycle_scoring}
\title{Cell cycle scoring}
\usage{
cell_cycle_scoring(input_read_RNA_assay, empty_droplets_tbl)
}
\description{
Applies cell cycle scoring based on the expression of G2/M and S phase markers. 
Returns a tibble containing cell identifiers with their predicted classification into 
cell cycle phase: G2M, S, or G1 phase.
}
