% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellChatDB.human}
\alias{CellChatDB.human}
\title{CellChatDB.human database}
\format{
A list with multiple elements, each representing different parts of the signaling network.
}
\source{
CellChat::CellChatDB.human
}
\usage{
data(CellChatDB.human)
}
\description{
A curated human ligand–receptor interaction database provided by the CellChat package.
}
\details{
This object is typically used as input to the CellChat pipeline. It contains signaling pathway data
for cell-cell communication analysis.
}
\keyword{datasets}
