% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tranform_assay.R
\name{census_harmonise_anndata_counts}
\alias{census_harmonise_anndata_counts}
\title{Harmonize Counts Data in a SingleCellExperiment Object}
\usage{
census_harmonise_anndata_counts(data, transform_method)
}
\arguments{
\item{data}{A \code{SingleCellExperiment} object containing assays with counts data.}

\item{transform_method}{A function or the name of a function (as a character string) to transform the counts data (e.g., \code{"log1p"}, \code{"exp"}).}
}
\value{
The modified \code{SingleCellExperiment} object with harmonized counts.
}
\description{
This function harmonizes the counts data in a \code{SingleCellExperiment} object by adjusting negative values,
scaling counts to avoid downstream failures, applying a transformation method, and removing cells with zero counts.
}
\examples{
\dontrun{
library(SingleCellExperiment)
# Using a function object
transformed_sce <- census_harmonise_anndata_counts(sce, log1p)
# Using a function name as a character string
transformed_sce <- census_harmonise_anndata_counts(sce, "exp")
}

}
