% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean_cell_types}
\alias{clean_cell_types}
\title{Clean and Standardize Cell Types}
\usage{
clean_cell_types(.x)
}
\arguments{
\item{.x}{A vector of cell types.}
}
\value{
A cleaned and standardized vector of cell types.
}
\description{
This function takes a vector of cell types and applies a series of transformations
to clean and standardize them for better consistency.
}
\examples{
cell_types <- c("CD4+ T-cells", "NK cells", "Blast-cells")
}
