% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean_cell_types_deeper}
\alias{clean_cell_types_deeper}
\title{Clean and Standardize Cell Types (Deeper)}
\usage{
clean_cell_types_deeper(x)
}
\arguments{
\item{x}{A vector of cell types.}
}
\value{
A cleaned and standardized vector of cell types.
}
\description{
This function takes a vector of cell types and applies a series of transformations
to clean and standardize them for better consistency.
}
\examples{
cell_types <- c("CD4 T Cell, AlphaBeta", "NK cell, gammadelta", "Central Memory")
cleaned_cell_types <- clean_cell_types_deeper(cell_types)

}
