% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean_sce_metadata}
\alias{clean_sce_metadata}
\title{Clean Metadata in SingleCellExperiment Object}
\usage{
clean_sce_metadata(sce)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing metadata to be cleaned.}
}
\value{
A SingleCellExperiment with all completely NA columns removed
}
\description{
This function takes a SingleCellExperiment (SCE) object and removes columns
that are completely filled with NA values.
The cleaned metadata is then returned as a dataframe.
}
