% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_gene_names}
\alias{convert_gene_names}
\title{Gene name conversion using ensembl database}
\usage{
convert_gene_names(id, current_nomenclature)
}
\arguments{
\item{id}{Character vector of gene names}

\item{current_nomenclature}{Character vector of input gene nomenclature}
}
\value{
A data frame of gene name before and after conversion
}
\description{
Gene name conversion using ensembl database
}
