% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPCell.R
\name{eliminate_random_effects}
\alias{eliminate_random_effects}
\title{HPCell Package Functions}
\usage{
eliminate_random_effects(formula)
}
\arguments{
\item{formula}{An object of class \code{formula}, representing a mixed-effects model formula.}
}
\value{
A formula object with random effects parts removed.
}
\description{
Functions for the HPCell package.
Eliminate Random Effects from a Formula
}
\examples{
eliminate_random_effects(~ age_days * sex + (1 | file_id) + ethnicity_simplified + assay_simplified + .aggregated_cells + (1 + age_days * sex | tissue))
}
