% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ensembl_genes_biomart}
\alias{ensembl_genes_biomart}
\title{A data frame of Ensembl genes retrieved from biomaRt package}
\format{
A data frame map of ensembl_gene_id, external_gene_name and chromosome_name
}
\source{
biomaRt::getBM()
}
\usage{
data(ensembl_genes_biomart)
}
\description{
This dataset contains Ensembl gene IDs, external gene names, and chromosome names
retrieved using the biomaRt package.
}
\keyword{datasets}
