% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_curated_constructor.R
\name{ensemble_annotation}
\alias{ensemble_annotation}
\title{Ensemble Annotation for Cell Type Identification}
\usage{
ensemble_annotation(
  celltype_matrix,
  method_weights = NULL,
  override_celltype = c(),
  celltype_tree = NULL
)
}
\arguments{
\item{celltype_matrix}{A matrix or data frame where columns represent different annotation
methods for cell types. Each element in the matrix represents a cell type determined by
each method.}

\item{method_weights}{Optional numeric vector or matrix specifying weights for each method.
If not provided, equal weights are used. If provided as a vector, it should match the
number of methods (columns of celltype_matrix).}

\item{override_celltype}{A character vector of cell types that should override the voting
process if they appear. This can be used to set certain cell types as non-negotiable
when they are detected by any method.}

\item{celltype_tree}{An igraph object representing the hierarchy of cell types. If NULL,
a default graph named "immune_graph" from the global environment is used.}
}
\value{
A vector representing the consensus cell type for each row in the input \code{celltype_matrix}.
}
\description{
This function creates an ensemble annotation for cell types by utilizing a voting mechanism
across different methods. It leverages a hierarchy of cell types, method-specific weights,
and an option to override certain cell types to derive a consensus classification.
}
