% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{feature_chunks}
\alias{feature_chunks}
\title{Divide Features into Chunks}
\usage{
feature_chunks(features, chunk_size = 100)
}
\arguments{
\item{features}{A vector of features to be divided into chunks.}

\item{chunk_size}{The size of each chunk. Defaults to 100.}
}
\value{
A tibble with the features and their corresponding chunk numbers.
}
\description{
This function divides a list of features into chunks of a specified size.
}
