% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_variable_genes}
\alias{find_variable_genes}
\title{Find variable genes}
\usage{
find_variable_genes(input_seurat, empty_droplet)
}
\arguments{
\item{input_seurat}{Single Seurat object (Input data)}

\item{empty_droplet}{Single dataframe containing empty droplet filtering information}
}
\value{
A vector of variable gene names
}
\description{
Find variable genes
}
