% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_count_per_gene_df}
\alias{get_count_per_gene_df}
\title{Transform counts to continous data}
\usage{
get_count_per_gene_df(counts)
}
\arguments{
\item{counts}{A SummarizedExperiment object}
}
\description{
Transform counts to continous data
}
