% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_unique_tissues}
\alias{get_unique_tissues}
\title{Subsetting input dataset into a list of SingleCellExperiment or Seurat objects by pre-specified sample column tissue}
\usage{
get_unique_tissues(sce_obj, sample_column)
}
\arguments{
\item{sce_obj}{A \code{SingleCellExperiment} or \code{Seurat} object containing input single-cell data}

\item{sample_column}{The column name specifying sample information}
}
\value{
The unique sample types in the sample column

@description
Function to subset \code{SingleCellExperiment} or \code{Seurat} object by tissue
}
\description{
Subsetting input dataset into a list of SingleCellExperiment or Seurat objects by pre-specified sample column tissue
}
