% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{harmonise_names_non_immune}
\alias{harmonise_names_non_immune}
\title{Harmonize Non-Immune Cell Type Names}
\usage{
harmonise_names_non_immune(metadata)
}
\arguments{
\item{metadata}{A data frame containing cell type information.}
}
\value{
The metadata with harmonized cell type names.
}
\description{
This function harmonizes non-immune cell type names in the metadata.
}
\examples{
metadata <- data.frame(cell_type = c("Myofibroblast", "Fibroblast", "Other Fibroblast"))

}
