% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPCell.R
\docType{package}
\name{HPCell-package}
\alias{HPCell}
\alias{HPCell-package}
\title{HPCell: Massively-Parallel R Native Pipeline for Single-Cell Analysis}
\description{
Massively-parallel R native pipeline for single-cell analysis.
}
\author{
\strong{Maintainer}: Stefano Mangiola \email{mangiolastefano@gmail.com}

Authors:
\itemize{
  \item Jiayi Si \email{si.j@wehi.edu.au}
}

}
