% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_functions.R
\name{hpcell_map_test_differential_abundance}
\alias{hpcell_map_test_differential_abundance}
\title{Main Function for HPCell Map Test Differential Abundance}
\usage{
hpcell_map_test_differential_abundance(
  data_df,
  .formula_column,
  .data_column,
  .group_name_columns,
  .abundance = NULL,
  store = tempfile(tmpdir = "."),
  computing_resources = crew_controller_local(workers = 1),
  cpus_per_task = 1,
  debug_job_id = NULL,
  append = FALSE
)
}
\arguments{
\item{data_df}{Data frame to be processed.}

\item{.data_column}{Column in the data frame containing the data.}

\item{store}{File path for temporary storage.}

\item{computing_resources}{Computing resources configuration.}

\item{cpus_per_task}{Number of CPUs allocated per task.}

\item{debug_job_id}{Optional job ID for debugging.}

\item{append}{Flag to append to existing script.}

\item{formula}{Formula for the differential abundance test.}
}
\value{
A \code{targets} pipeline output, typically a nested tibble with differential abundance estimates.
}
\description{
This function prepares and runs a differential abundance test pipeline using the 'targets' package. It sets up necessary files, appends scripts, and executes the pipeline.
}
