% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map_add_dispersion_to_se}
\alias{map_add_dispersion_to_se}
\title{Add Dispersion Estimates to SingleCellExperiment Object}
\usage{
map_add_dispersion_to_se(se_df, .col, abundance = NULL)
}
\arguments{
\item{se_df}{A data frame or list containing SingleCellExperiment objects.}

\item{.col}{A symbol indicating the column in \code{se_df} that contains SingleCellExperiment objects.}

\item{abundance}{(Optional) A character vector specifying the name of the assay to be used
for dispersion estimation. If NULL or not provided, the first assay is used.}
}
\value{
The input data frame or list (\code{se_df}) with the specified \code{.col} modified to include
dispersion estimates in each SingleCellExperiment object.
}
\description{
\code{map_add_dispersion_to_se} function adds dispersion estimates to each feature (gene) in a
SingleCellExperiment object. Dispersion estimates are added based on the abundance measure specified.
}
\details{
The function iterates over each SingleCellExperiment object in the specified column of the input data frame
or list. It calculates dispersion estimates for the features (genes) based on the specified abundance assay.
The results are joined back to each SingleCellExperiment object. If no abundance assay is specified,
the function defaults to the first assay in each SingleCellExperiment object.
}
