% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map_split_sce_by_gene}
\alias{map_split_sce_by_gene}
\title{map_split_sce_by_gene Split SingleCellExperiment by Gene}
\usage{
map_split_sce_by_gene(
  sce_df,
  .col,
  how_many_chunks_base = 10,
  max_cells_before_split = 4763
)
}
\arguments{
\item{sce_df}{A dataframe (or tibble) where one of the columns contains SingleCellExperiment objects}

\item{.col}{A symbol or string indicating the column in \code{sce_df} which should be dynamically split into multiple chunks}

\item{how_many_chunks_base}{A base number of chunks to divide the data into, adjusted by the actual size of the data in each group.}

\item{max_cells_before_split}{The maximum number of cells a single chunk can have before it is split into another chunk.}
}
\value{
Returns the input SingleCellExperiment DataFrame with an additional column \code{sce_md5} containing MD5 hashes of the chunks, and with the data split according to the specified parameters.
}
\description{
Splits a SingleCellExperiment object into multiple chunks based on the number of cells.
This function dynamically partitions a SingleCellExperiment object into multiple chunks based on the number of cells per gene across the specified column. It computes the number of splits by dividing the total number of cells by a maximum threshold and multiplying the result by a base number of chunks. This approach allows handling of large datasets by reducing the complexity in each chunk, making it feasible to perform detailed analyses or computational tasks on subsets of data efficiently.
The function also assigns a unique MD5 hash to each chunk as an identifier, facilitating tracking and referencing of data subsets in subsequent analyses.
}
