% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map_split_se_by_gene}
\alias{map_split_se_by_gene}
\title{Split SummarizedExperiment Object by Gene}
\usage{
map_split_se_by_gene(se_df, .col, .number_of_chunks)
}
\arguments{
\item{se_df}{Data frame containing SummarizedExperiment objects.}

\item{.col}{Column in the data frame containing the SummarizedExperiment objects.}

\item{.number_of_chunks}{Number of chunks to split into.}
}
\value{
Data frame with SummarizedExperiment objects split into chunks.
}
\description{
Splits each SummarizedExperiment object in a data frame into chunks by gene.
}
