% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map_split_se_by_number_of_genes}
\alias{map_split_se_by_number_of_genes}
\title{map_split_se_by_number_of_genes}
\usage{
map_split_se_by_number_of_genes(se_df, .col, chunk_size = 100)
}
\arguments{
\item{se_df}{A \code{SingleCellExperiment} DataFrame that contains gene expression or other related data.}

\item{.col}{A symbol or string indicating the column in \code{se_df} which should be dynamically split into multiple chunks based on gene count.}

\item{chunk_size}{The maximum number of genes each chunk should contain; defaults to 100.}
}
\value{
Returns the input \code{SingleCellExperiment} DataFrame with an additional column \code{se_md5} containing unique random IDs for each chunk, and with the data split into chunks according to the specified number of genes.
}
\description{
Splits the singlecellexperiment dataframe into multiple chunks based on gene count
}
