% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonimmune_cellxgene}
\alias{nonimmune_cellxgene}
\title{Non-immune CellxGene cell-type labels}
\format{
A character vector.
}
\usage{
data(nonimmune_cellxgene)
}
\description{
A character vector of cell-type labels from the CellxGene corpus that are
classified as non-immune. Used by \code{cell_type_ensembl_harmonised()} to
distinguish immune from non-immune populations during consensus annotation.
}
\keyword{datasets}
