% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{postprocess_SCimplify}
\alias{postprocess_SCimplify}
\title{Detection of metacells with the SuperCell approach}
\usage{
postprocess_SCimplify(
  preprocessed,
  cell.annotation = NULL,
  cell.split.condition = NULL,
  gamma,
  block.size = 10000,
  igraph.clustering = c("walktrap", "louvain"),
  return.singlecell.NW = TRUE,
  return.hierarchical.structure = TRUE,
  ...
)
}
\arguments{
\item{preprocessed}{A list returned by \code{preprocess_SCimplify} containing preprocessed single-cell data,
PCA results, and kNN graph.}

\item{cell.annotation}{a vector of cell type annotation, if provided, metacells that contain single cells of different cell type annotation will be split in multiple pure metacell (may result in slightly larger numbe of metacells than expected with a given gamma)}

\item{cell.split.condition}{a vector of cell conditions that must not be mixed in one metacell. If provided, metacells will be split in condition-pure metacell (may result in significantly(!) larger number of metacells than expected)}

\item{gamma}{graining level of data (proportion of number of single cells in the initial dataset to the number of metacells in the final dataset)}

\item{block.size}{number of cells to map to the nearest metacell at the time (for approx coarse-graining)}

\item{igraph.clustering}{clustering method to identify metacells (available methods "walktrap" (default) and "louvain" (not recommended, gamma is ignored)).}

\item{return.singlecell.NW}{whether return single-cell network (which consists of approx.N if \code{"do.approx"} or all cells otherwise)}

\item{return.hierarchical.structure}{whether return hierarchical structure of metacell}

\item{...}{other parameters of \link{build_knn_graph} function}
}
\value{
A tibble with column 'cell' and 'membership' indicating which metacell cluster each cell belongs to.
}
\description{
This function detects metacells (former super-cells) from single-cell gene expression matrix
}
