% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{preprocess_SCimplify}
\alias{preprocess_SCimplify}
\title{Preprocess metacells with the SuperCell approach}
\usage{
preprocess_SCimplify(
  input_read_RNA_assay,
  assay = NULL,
  genes.use = NULL,
  genes.exclude = NULL,
  n.var.genes = min(1000, nrow(input_read_RNA_assay)),
  k.knn = 5,
  do.scale = TRUE,
  n.pc = 10,
  fast.pca = TRUE,
  do.approx = FALSE,
  approx.N = 5000,
  seed = 12345,
  ...
)
}
\arguments{
\item{input_read_RNA_assay}{A \code{SingleCellExperiment} or \code{Seurat} object containing RNA assay data.}

\item{assay}{assay used, default = "RNA"}

\item{genes.use}{a vector of genes used to compute PCA}

\item{genes.exclude}{a vector of genes to be excluded when computing PCA}

\item{n.var.genes}{if \code{"genes.use"} is not provided, \code{"n.var.genes"} genes with the largest variation are used}

\item{k.knn}{parameter to compute single-cell kNN network}

\item{do.scale}{whether to scale gene expression matrix when computing PCA}

\item{n.pc}{number of principal components to use for construction of single-cell kNN network}

\item{fast.pca}{use \link[irlba]{irlba} as a faster version of prcomp (one used in Seurat package)}

\item{do.approx}{compute approximate kNN in case of a large dataset (>50'000)}

\item{approx.N}{number of cells to subsample for an approximate approach. By default, 5000 cells are used
for approximation to capture biological meaningful result.}

\item{seed}{seed to use to subsample cells for an approximate approach}

\item{...}{other parameters of \link{build_knn_graph} function}
}
\value{
A list of variables to be passed to the \code{SuperCell::SCimplify} gamma involved function.
}
\description{
This function preprocesses a single-cell gene expression matrix for downstream simplification using PCA
and k-nearest neighbor (kNN) graph construction. It includes options for scaling, feature selection,
approximate sampling, and PCA computation methods.
}
