% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pseudobulk_merge}
\alias{pseudobulk_merge}
\title{Merge pseudobulk from all samples}
\usage{
pseudobulk_merge(pseudobulk_list, external_path, ...)
}
\arguments{
\item{pseudobulk_list}{A list of pseudobulk samples generated by \code{create_pseudobulk()}.}

\item{external_path}{Character path to the directory where the merged HDF5
pseudobulk file is written.}

\item{...}{Additional arguments passed to internal functions.}
}
\description{
Merge pseudobulk from all samples. Ensures that missing genes are accounted
for and aligns data across multiple samples.
}
