% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pseudobulk_preprocessing}
\alias{pseudobulk_preprocessing}
\title{Pseudobulk_preprocessing}
\usage{
pseudobulk_preprocessing(
  reference_label_fine,
  preprocessing_output_S,
  sample_column
)
}
\description{
Pseudobulk_preprocessing aggregates cells based on sample and cell type annotations, 
creating pseudobulk samples for each combination. It handles RNA and ADT assays, 
and ensures that missing genes are accounted for and aligns data across multiple samples. 
Returns a list containing pseudobulk data aggregated by sample and by both 
sample and cell type if the reference label for cell type is present.
}
\details{
This function input a list of Seurat objects and outputs a unique Summa...
}
