% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reference_annotation_to_consensus}
\alias{reference_annotation_to_consensus}
\title{reference_annotation_to_consensus}
\usage{
reference_annotation_to_consensus(azimuth_input, monaco_input, blueprint_input)
}
\arguments{
\item{azimuth_input}{A vector of cell type annotations from the Azimuth dataset.}

\item{monaco_input}{A vector of cell type annotations from the Monaco dataset.}

\item{blueprint_input}{A vector of cell type annotations from the Blueprint dataset.}
}
\value{
A vector of consensus cell type annotations, merging inputs from the three datasets.
}
\description{
This function takes cell type annotations from multiple datasets (Azimuth, Monaco, Blueprint) and harmonizes them into a consensus annotation. The function utilizes predefined mappings between cell type labels in these datasets to generate standardized cell types across references.
}
\note{
This function is designed to harmonize specific cell types, especially T cells, B cells, monocytic cells, and innate lymphoid cells (ILCs), across reference datasets.
}
\examples{
# Example usage:
tibble::tibble(
  azimuth_predicted.celltype.l2 = c("CD8 TEM", "NK", "CD4 Naive"),
  monaco_first.labels.fine = c("Effector memory CD8 T cells", "Natural killer cells", "Naive CD4 T cells"),
  blueprint_first.labels.fine = c("CD8+ Tem", "NK cells", "Naive B-cells")
) |>
  dplyr::mutate(consensus = reference_annotation_to_consensus(
    azimuth_predicted.celltype.l2, monaco_first.labels.fine, blueprint_first.labels.fine))

}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[stringr]{str_detect}}, \code{\link[tidyr]{expand_grid}}
}
