% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reference_label_fine_id}
\alias{reference_label_fine_id}
\title{Reference_label_fine}
\usage{
reference_label_fine_id(tissue)
}
\description{
Add reference label depending on the provided tissue type of the input data
}
