% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_pipeline.R
\name{run_targets_pipeline}
\alias{run_targets_pipeline}
\title{Run Targets Pipeline for HPCell}
\usage{
run_targets_pipeline(
  input_data,
  store = "./",
  input_reference = NULL,
  tissue,
  computing_resources = crew_controller_local(workers = 1),
  debug_step = NULL,
  filter_empty_droplets = TRUE,
  RNA_assay_name = "RNA",
  sample_column = "sample",
  my_launcher = custom_launcher_class$new(name = "my_custom_launcher", seconds_interval =
    10, seconds_timeout = 300, seconds_launch = 5, seconds_idle = 60, seconds_wall =
    3600, tasks_max = 20, tasks_timers = 45, launch_max = 25, reset_globals = FALSE,
    reset_packages = FALSE, reset_options = FALSE, garbage_collection = TRUE, tls =
    crew_tls())
)
}
\arguments{
\item{input_data}{Input data for the pipeline.}

\item{store}{Directory path for storing the pipeline files.}

\item{input_reference}{Optional reference data.}

\item{tissue}{Tissue type for the analysis.}

\item{computing_resources}{Configuration for computing resources.}

\item{debug_step}{Optional step for debugging.}

\item{filter_empty_droplets}{Flag to indicate if input filtering is needed.}

\item{RNA_assay_name}{Name of the RNA assay.}

\item{sample_column}{Column name for sample identification.}
}
\value{
The output of the \code{targets} pipeline, typically a preprocessed dataset.
}
\description{
This function sets up and executes a \code{targets} pipeline for HPCell. It saves input data and configurations,
writes a pipeline script, and runs the pipeline using the 'targets' package.
}
