% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{save_experiment_data}
\alias{save_experiment_data}
\title{Save various types of single-cell data}
\usage{
save_experiment_data(data, dir, container_type = "anndata")
}
\arguments{
\item{data}{A data object to save.}

\item{dir}{A character vector of length one specifies the file path, or directory path.}

\item{container_type}{A character vector of length one specifies the input data type.}
}
\value{
An object stored in the defined path.
}
\description{
Save various types of single-cell data
}
