% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sct_v2_trycatch_fallback}
\alias{sct_v2_trycatch_fallback}
\title{Run SCTransform(v2) fallback hack after the "incorrect number of dimensions" crash}
\usage{
sct_v2_trycatch_fallback(
  obj,
  assay = "X",
  factors_to_regress = NULL,
  scale_factor = 2186,
  conserve.memory = TRUE,
  min_cells = 0,
  return.only.var.genes = FALSE,
  residual.features = NULL,
  verbose = TRUE,
  min_cells_for_survivor_detection = 5,
  host_pick_top_detected = 2000,
  jitter_range = c(0.99, 1.01),
  ...
)
}
\value{
A Seurat object returned by \code{Seurat::SCTransform()},
with provenance attributes describing the fallback.
}
\description{
This function is intended to be called \emph{only after} \code{Seurat::SCTransform()}
with \code{vst.flavor = "v2"} fails with the error
\emph{"Error in model_pars\link{, 'theta'} : incorrect number of dimensions"}.
}
