% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{se_add_dispersion}
\alias{se_add_dispersion}
\title{Add Dispersion Estimates to SummarizedExperiment}
\usage{
se_add_dispersion(se, my_formula, my_assay)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{my_formula}{A formula used to model the design matrix.}

\item{my_assay}{The assay to be used for dispersion estimation.}

\item{se_df}{A data frame containing SingleCellExperiment objects.}
}
\value{
The input data frame with updated SingleCellExperiment objects containing dispersion values.
}
\description{
Adds dispersion estimates to a SummarizedExperiment object using edgeR.
For smaller datasets (<1000 samples) all samples are used; for larger datasets
a random subset of up to 2000 samples is drawn to reduce computational load.
}
