% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellChat.R
\name{seurat_to_ligand_receptor_count}
\alias{seurat_to_ligand_receptor_count}
\title{Ligand-Receptor Count from Seurat Data}
\usage{
seurat_to_ligand_receptor_count(
  counts,
  .cell_group,
  assay,
  sample_for_plotting = ""
)
}
\arguments{
\item{counts}{Seurat object.}

\item{.cell_group}{Cell group variable.}

\item{assay}{Name of the assay to use.}

\item{sample_for_plotting}{Sample name for plotting.}
}
\value{
A list of communication results including interactions and signaling pathways.
}
\description{
Calculates ligand-receptor interactions for each cell type in a Seurat object using CellChat.
}
