% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{split_sample_cell_type_calculate_metacell_membership}
\alias{split_sample_cell_type_calculate_metacell_membership}
\title{Calculate Metacell Membership for Each Cell Type}
\usage{
split_sample_cell_type_calculate_metacell_membership(
  sample_sce,
  cell_type_tbl,
  x = "cell_type",
  min_cells_per_metacell = 30
)
}
\arguments{
\item{sample_sce}{A SingleCellExperiment object containing single-cell data.}

\item{cell_type_tbl}{A tibble of cell type.}

\item{x}{A character vector of cell type aggregation column.}

\item{min_cells_per_metacell}{An integer of minimum cells in each metacell.}
}
\value{
A tibble with metacell membership data for each cell type.
}
\description{
This function processes a SingleCellExperiment object by grouping cells according
to their type, calculates metacell membership for each group, and combines the
results into a single tibble.
}
