% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{split_summarized_experiment}
\alias{split_summarized_experiment}
\title{Split a SummarizedExperiment object into chunks of specified size}
\usage{
split_summarized_experiment(se, chunk_size = 100)
}
\arguments{
\item{se}{A SummarizedExperiment object to be split.}

\item{chunk_size}{The number of genes per chunk (default is 100).}
}
\value{
A list of SummarizedExperiment objects, each containing a chunk of genes.
}
\description{
This function splits a SummarizedExperiment object into smaller chunks,
each containing a specified number of genes.
}
\examples{
# Assume `se` is your SummarizedExperiment object
# se <- your SummarizedExperiment object here
# Split the SummarizedExperiment object into chunks of 100 genes each
# se_chunks <- split_summarized_experiment(se)
}
