% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{theme_multipanel}
\alias{theme_multipanel}
\title{theme_multipanel}
\format{
A ggplot theme
}
\usage{
data(theme_multipanel)
}
\description{
this object includes the ggplot, theme
}
\keyword{internal}
