% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{vector_to_code}
\alias{vector_to_code}
\title{Convert a vector of integers to its R code form as a character}
\usage{
vector_to_code(int_vector)
}
\arguments{
\item{int_vector}{A vector of integers.}
}
\value{
A character string representing the R code form of the vector.
}
\description{
This function takes a vector of integers and returns its R code form as a character string.
}
\examples{
int_vector <- c(1, 2, 3, 4, 5)
code_string <- vector_to_code(int_vector)
print(code_string)
}
