% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alive_identification}
\alias{alive_identification}
\title{alive_identification}
\usage{
alive_identification(
  input_read_RNA_assay,
  empty_droplets_tbl,
  annotation_label_transfer_tbl
)
}
\description{
Filters out dead cells by analyzing mitochondrial and ribosomal gene expression 
percentages
}
