% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alive_identification}
\alias{alive_identification}
\title{Alive Cell Identification}
\usage{
alive_identification(
  input_read_RNA_assay,
  empty_droplets_tbl,
  annotation_label_transfer_tbl,
  assay = NULL,
  tissue_name
)
}
\arguments{
\item{input_read_RNA_assay}{SingleCellExperiment object containing RNA assay data.}

\item{empty_droplets_tbl}{A tibble identifying empty droplets.}

\item{annotation_label_transfer_tbl}{A tibble with annotation label transfer data.}

\item{assay}{The assay to be used for analysis, specified as a character string.}
}
\value{
A tibble identifying alive cells.
}
\description{
\code{alive_identification} filters out dead cells by analyzing mitochondrial and ribosomal gene expression percentages.
}
