% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{annotation_label_transfer}
\alias{annotation_label_transfer}
\title{annotation_label_transfer}
\usage{
annotation_label_transfer(
  input_read_RNA_assay,
  reference_azimuth,
  empty_droplets_tbl
)
}
\description{
annotation_label_transfer employs SingleR for cell-type identification using
reference datasets (Blueprint and Monaco Immune data). It also performs cell type labelling 
using Azimuth when a reference is provided.
}
