% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{annotation_label_transfer}
\alias{annotation_label_transfer}
\title{Cell Type Annotation Transfer}
\usage{
annotation_label_transfer(
  input_read_RNA_assay,
  empty_droplets_tbl = NULL,
  reference_azimuth = NULL,
  assay = NULL,
  feature_nomenclature
)
}
\arguments{
\item{input_read_RNA_assay}{A \code{SingleCellExperiment} or \code{Seurat} object containing RNA assay data.}

\item{empty_droplets_tbl}{A tibble identifying empty droplets.}

\item{reference_azimuth}{Optional reference data for Azimuth.}

\item{assay}{assay used, default = "RNA"}
}
\value{
A tibble with cell type annotation data.
}
\description{
\code{annotation_label_transfer} utilizes SingleR for cell-type identification using reference datasets
(Blueprint and Monaco Immune data). It can also perform cell type labeling using Azimuth when a reference
is provided.
}
