% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_UMAP}
\alias{calc_UMAP}
\title{Calculate UMAP
Scales the input data, performing PCA, clustering the cells and running UMAP and constructs a tibble in preparation for plotting in
the doublet identification report}
\usage{
calc_UMAP(input_seurat)
}
\arguments{
\item{input_seurat}{Single Seurat object (Input data)}
}
\description{
Calculate UMAP
Scales the input data, performing PCA, clustering the cells and running UMAP and constructs a tibble in preparation for plotting in
the doublet identification report
}
