% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_gamma}
\alias{calculate_gamma}
\title{Calculate Appropriate Gamma Values for Metacell Analysis}
\usage{
calculate_gamma(cell_count, min_cells_per_metacell = 30)
}
\arguments{
\item{cell_count}{Integer, the total number of cells.}

\item{min_cells_per_metacell}{Integer, the minimum number of cells per metacell. Defaults to 30.}
}
\value{
An Integer vector of viable gamma values. If no viable gamma values are found, returns 0.
}
\description{
This function determines viable gamma values to be used in metacell analysis. It calculates gamma values
by doubling gamma until the quotient of the total cell count and gamma is less than the specified minimum
number of cells per metacell.
}
