% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cluster_metacell}
\alias{cluster_metacell}
\title{Metacell Clustering}
\usage{
cluster_metacell(
  input_hpc,
  target_input = "data_object",
  target_output = "metacell_classification_tbl",
  ...
)
}
\arguments{
\item{input_read_RNA_assay}{A \code{SingleCellExperiment} or \code{Seurat} object containing RNA assay data.}

\item{empty_droplets_tbl}{A tibble identifying empty droplets.}

\item{alive_identification_tbl}{A tibble from alive cell identification.}

\item{cell_cycle_score_tbl}{A tibble from cell cycle scoring.}

\item{assay}{assay used, default = "RNA"}
}
\value{
A tibble with column 'cell' and 'membership' indicating which metacell cluster each cell belongs to.
}
\description{
This function processes single-cell RNA sequencing data to cluster cells into metacells,
a higher resolution of clustering that groups cells sharing similar gene expression patterns.
}
