% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_pseudobulk}
\alias{create_pseudobulk}
\title{Create pseudobulk}
\usage{
create_pseudobulk(preprocessing_output_S, assays, x, ...)
}
\arguments{
\item{preprocessing_output_S}{Processed dataset from preprocessing.}

\item{assays}{assay used, default = "RNA"}
}
\value{
List containing pseudobulk data aggregated by sample and by both sample and cell type.
}
\description{
Aggregates cells based on sample and cell type annotations, creating pseudobulk samples
for each combination. Handles RNA and ADT assays
}
