% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{doublet_identification}
\alias{doublet_identification}
\title{Doublet Identification}
\usage{
doublet_identification(
  input_read_RNA_assay,
  empty_droplets_tbl,
  alive_identification_tbl,
  annotation_label_transfer_tbl,
  reference_label_fine
)
}
\arguments{
\item{input_read_RNA_assay}{SingleCellExperiment object containing RNA assay data.}

\item{empty_droplets_tbl}{A tibble identifying empty droplets.}

\item{alive_identification_tbl}{A tibble identifying alive cells.}

\item{annotation_label_transfer_tbl}{A tibble with annotation label transfer data.}

\item{reference_label_fine}{Optional reference label for fine-tuning.}
}
\value{
A tibble containing cells with their scDblFinder scores.
}
\description{
\code{doublet_identification} applies the scDblFinder algorithm to the filtered dataset. It supports integrating with
SingleR annotations if provided and outputs a tibble containing cells with their associated scDblFinder scores.
}
