% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{doublet_identification}
\alias{doublet_identification}
\title{Doublet identification}
\usage{
doublet_identification(
  input_read_RNA_assay,
  empty_droplets_tbl,
  alive_identification_tbl,
  annotation_label_transfer_tbl,
  reference_label_fine
)
}
\description{
applies the scDblFinder algorithm to the filtered dataset. It supports integrating with 
SingleR annotations if provided. Outputs a tibble containing cells with their associated 
scDblFinder scores, indicating the likelihood of being a doublet.
}
