% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{empty_droplet_id}
\alias{empty_droplet_id}
\title{empty_droplet_id}
\usage{
empty_droplet_id(input_read_RNA_assay, filter_input)
}
\description{
empty_droplet_id uses the DropletUtils package to apply a significance threshold to differentiate between empty and 
non-empty droplets. The function excludes mitochondrial and ribosomal genes,calculates barcode ranks and optionally
filters the input data based on these criteria. The outputs a tibble with logprob, FDR and classification whether the
cells are empty droplets
}
