% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{empty_droplet_id}
\alias{empty_droplet_id}
\title{Identify Empty Droplets in Single-Cell RNA-seq Data}
\usage{
empty_droplet_id(
  input_read_RNA_assay,
  total_RNA_count_check = -Inf,
  assay = NULL,
  feature_nomenclature
)
}
\arguments{
\item{input_read_RNA_assay}{SingleCellExperiment or Seurat object containing RNA assay data.}

\item{filter_empty_droplets}{Logical value indicating whether to filter the input data.}
}
\value{
A tibble with columns: logProb, FDR, empty_droplet (classification of droplets).
}
\description{
\code{empty_droplet_id} distinguishes between empty and non-empty droplets using the DropletUtils package.
It excludes mitochondrial and ribosomal genes, calculates barcode ranks, and optionally filters input data
based on these criteria. The function returns a tibble containing log probabilities, FDR, and a classification
indicating whether cells are empty droplets.
}
