% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{empty_droplet_threshold}
\alias{empty_droplet_threshold}
\title{Identify Empty Droplets in Single-Cell RNA-seq Data}
\usage{
empty_droplet_threshold(
  input_read_RNA_assay,
  total_RNA_count_check = -Inf,
  assay = NULL,
  gene_nomenclature,
  RNA_count_threshold = 100,
  RNA_feature_threshold = 200
)
}
\arguments{
\item{input_read_RNA_assay}{SingleCellExperiment or Seurat object containing RNA assay data.}

\item{RNA_count_threshold}{An optional integer for the number of RNA count per cell. Default uses 100}

\item{RNA_feature_threshold}{An optional integer for the number of feature count. Default uses 200}

\item{filter_empty_droplets}{Logical value indicating whether to filter the input data.}
}
\value{
A tibble with columns: Cell, nFeature_RNA, nCount_RNA, empty_droplet (classification of droplets).
}
\description{
\code{empty_droplet_threshold} distinguishes between empty and non-empty droplets by threshold.
It excludes mitochondrial and ribosomal genes, and filters input data
based on defined values of \code{nCount_RNA} and \code{nFeature_RNA}
The function returns a tibble containing RNA count, RNA feature count indicating whether cells are empty droplets.
}
