% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{empty_droplet_threshold}
\alias{empty_droplet_threshold}
\title{Identify Empty Droplets in Single-Cell RNA-seq Data}
\usage{
empty_droplet_threshold(
  input_read_RNA_assay,
  total_RNA_count_check = -Inf,
  assay = NULL,
  feature_nomenclature,
  RNA_feature_threshold = NULL
)
}
\arguments{
\item{input_read_RNA_assay}{SingleCellExperiment or Seurat object containing RNA assay data.}

\item{RNA_feature_threshold}{An optional integer for the number of feature expressed in a sample.}

\item{filter_empty_droplets}{Logical value indicating whether to filter the input data.}
}
\value{
A tibble with columns: Cell, nFeature_expressed_in_sample, nCount_RNA, empty_droplet (classification of droplets).
}
\description{
\code{empty_droplet_threshold} identifies empty droplets by applying a gene expression threshold per sample.
It excludes mitochondrial and ribosomal genes, and classifies droplets as empty if
the number of expressed genes falls below the specified threshold.

The function returns a tibble containing the number of expressed genes,
total RNA count for each cell, and a logical annotation indicating whether the droplet was classified as empty.
}
