% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/functions.R
=======
% Please edit documentation in R/utilities.R
>>>>>>> f83a64c1aec4e358a94be07aa04d967a5dc6e784
\name{get_unique_tissues}
\alias{get_unique_tissues}
\title{Subsetting input dataset into a list of seurat objects by sample/ tissue}
\usage{
get_unique_tissues(seurat_object)
}
\description{
Function to subset Seurat object by tissue
}
