% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_unique_tissues}
\alias{get_unique_tissues}
\title{Subsetting input dataset into a list of seurat objects by sample/ tissue}
\usage{
get_unique_tissues(seurat_object, sample_column)
}
\arguments{
\item{seurat_object}{A single seurat object (input data)}

\item{sample_column}{A symbol or string that specifies the column in the metadata from which to extract unique values

@description
Function to subset Seurat object by tissue}
}
\description{
Subsetting input dataset into a list of seurat objects by sample/ tissue
}
