% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_unique_tissues}
\alias{get_unique_tissues}
\title{Get unique tissues
Obtain unique tissues/ samples from input dataset}
\usage{
get_unique_tissues(seurat_object)
}
\description{
Get unique tissues
Obtain unique tissues/ samples from input dataset
}
