% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/utilities.R
\name{get_unique_tissues}
\alias{get_unique_tissues}
\title{Get unique tissues
Obtain unique tissues/ samples from input dataset}
\usage{
get_unique_tissues(seurat_object, sample_column)

get_unique_tissues(seurat_object, sample_column)
}
\description{
Get unique tissues
Obtain unique tissues/ samples from input dataset

Subsetting input dataset into a list of seurat objects by sample/ tissue
@description
Function to subset Seurat object by tissue
}
