% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_unique_tissues}
\alias{get_unique_tissues}
\title{Subsetting input dataset into a list of seurat objects by sample/ tissue}
\usage{
get_unique_tissues(seurat_object)
}
\description{
Function to subset Seurat object by tissue
}
