% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories.R
\name{hpc_merge}
\alias{hpc_merge}
\title{Add HPC step to pipeline}
\usage{
hpc_merge(
  input_hpc,
  target_output = NULL,
  user_function = NULL,
  user_function_source_path = NULL,
  ...
)
}
\arguments{
\item{input_hpc}{The input HPC object.}

\item{target_output}{The output target name (default: NULL).}

\item{user_function}{A custom function provided by the user (default: NULL).}

\item{...}{Additional arguments to pass to the internal functions.}
}
\description{
This function adds a new step to the HPC pipeline by appending the appropriate
targets to the target script. It allows the user to specify the input and output
targets, as well as a custom user function to be applied.
}
