% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factories.R
\name{hpc_report}
\alias{hpc_report}
\title{Add HPC step to pipeline}
\usage{
hpc_report(input_hpc, target_output = NULL, rmd_path = NULL, ...)
}
\arguments{
\item{input_hpc}{The input HPC object.}

\item{target_output}{The output target name (default: NULL).}

\item{...}{Additional arguments to pass to the internal functions.}

\item{user_function}{A custom function provided by the user (default: NULL).}
}
\description{
This function adds a new step to the HPC pipeline by appending the appropriate
targets to the target script. It allows the user to specify the input and output
targets, as well as a custom user function to be applied.
}
