% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_grammar_hpc.R
\name{initialise_hpc}
\alias{initialise_hpc}
\title{Run Targets Pipeline for HPCell}
\usage{
initialise_hpc(
  input_hpc,
  store = targets::tar_config_get("store"),
  computing_resources = crew_controller_local(workers = 1),
  tier = rep(1, length(input_hpc)),
  debug_step = NULL,
  RNA_assay_name = "RNA"
)
}
\arguments{
\item{input_hpc}{Input data for the pipeline.}

\item{store}{Directory path for storing the pipeline files.}

\item{computing_resources}{Configuration for computing resources.}

\item{debug_step}{Optional step for debugging.}

\item{RNA_assay_name}{Name of the RNA assay.}

\item{input_reference}{Optional reference data.}

\item{tissue}{Tissue type for the analysis.}

\item{filter_empty_droplets}{Flag to indicate if input filtering is needed.}

\item{sample_column}{Column name for sample identification.}

\item{cell_type_annotation_column}{Column name for cell type annotation in input data}
}
\value{
The output of the \code{targets} pipeline, typically a pre-processed data set.
}
\description{
This function sets up and executes a \code{targets} pipeline for HPCell. It saves input data and configurations,
writes a pipeline script, and runs the pipeline using the 'targets' package.
}
