% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_grammar_hpc.R
\name{initialise_hpc}
\alias{initialise_hpc}
\title{Run Targets Pipeline for HPCell}
\usage{
initialise_hpc(
  input_hpc,
  store = targets::tar_config_get("store"),
  computing_resources = crew_controller_local(workers = 1),
  tier = rep(1, length(input_hpc)),
  debug_step = NULL,
  RNA_assay_name = "RNA",
  gene_nomenclature = "symbol",
  data_container_type,
  verbosity = targets::tar_config_get("reporter_make"),
  error = NULL,
  update = "thorough",
  garbage_collection = 0,
  workspace_on_error = FALSE
)
}
\arguments{
\item{input_hpc}{Character vector of input data path for the pipeline.}

\item{store}{Directory path for storing the pipeline files.}

\item{computing_resources}{Configuration for computing resources.}

\item{debug_step}{Optional step for debugging.}

\item{RNA_assay_name}{Name of the RNA assay.}

\item{gene_nomenclature}{Character vector indicating gene nomenclature in input_data}

\item{data_container_type}{A character vector of length one specifies the input data type.
The accepted input data type are:
sce_rds for \code{SingleCellExperiment} RDS,
seurat_rds for \code{Seurat} RDS,
sce_hdf5 for \code{SingleCellExperiment} HDF5-based object
seurat_h5 for \code{Seurat} HDF5-based object}

\item{input_reference}{Optional reference data.}

\item{tissue}{Tissue type for the analysis.}

\item{filter_empty_droplets}{Flag to indicate if input filtering is needed.}

\item{sample_column}{Column name for sample identification.}

\item{cell_type_annotation_column}{Column name for cell type annotation in input data}
}
\value{
The output of the \code{targets} pipeline, typically a pre-processed data set.
}
\description{
This function sets up and executes a \code{targets} pipeline for HPCell. It saves input data and configurations,
writes a pipeline script, and runs the pipeline using the 'targets' package.
}
