% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_functions.R
\name{map2_test_differential_abundance_hpc}
\alias{map2_test_differential_abundance_hpc}
\title{Main Function for HPCell Map Test Differential Abundance}
\usage{
map2_test_differential_abundance_hpc(
  data_list,
  formula_list,
  .abundance = NULL,
  store = tempfile(tmpdir = "."),
  computing_resources = crew_controller_local(workers = 1),
  cpus_per_task = 1,
  debug_job_id = NULL,
  append = FALSE,
  ...
)
}
\arguments{
\item{data_list}{A list of datasets to be processed}

\item{formula_list}{A list of formulae corresponding to each dataset in \code{data_list}}

\item{store}{File path for temporary storage.}

\item{computing_resources}{Computing resources configuration.}

\item{cpus_per_task}{Number of CPUs allocated per task.}

\item{debug_job_id}{Optional job ID for debugging.}

\item{append}{Flag to append to existing script.}

\item{...}{Additional arguments}

\item{data_df}{Data frame to be processed.}

\item{formula}{Formula for the differential abundance test.}

\item{.data_column}{Column in the data frame containing the data.}

\item{abundance}{Optional; a symbol indicating the column that specifies the abundance measure. If NULL, the first assay name is used.}
}
\value{
A \code{targets} pipeline output, typically a nested tibble with differential abundance estimates.
}
\description{
This function prepares and runs a differential abundance test pipeline using the 'targets' package. It sets up necessary files, appends scripts, and executes the pipeline.
}
