% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_functions.R
\name{map2_test_differential_abundance_hpc}
\alias{map2_test_differential_abundance_hpc}
\title{Main Function for HPCell Map Test Differential Abundance}
\usage{
map2_test_differential_abundance_hpc(
  data_list,
  formula_list,
  .abundance = NULL,
  store = tempfile(tmpdir = "."),
  computing_resources = crew_controller_local(workers = 1),
  cpus_per_task = 1,
  debug_job_id = NULL,
  append = FALSE,
  ...
)
}
\arguments{
\item{data_list}{list of dataframes to be processed}

\item{formula_list}{List of formula for the differential abundance test.}

\item{.abundance}{(optional) A symbol or string indicating the column name in the \code{SingleCellExperiment} object to be used for abundance measures. If not explicitly provided, the function attempts to automatically detect an appropriate column by examining the first object in \code{data_list}.}

\item{store}{File path for temporary storage.}

\item{computing_resources}{Computing resources configuration.}

\item{cpus_per_task}{Number of CPUs allocated per task.}

\item{debug_job_id}{Optional job ID for debugging.}

\item{append}{Flag to append to existing script.}

\item{...}{additional arguments}
}
\value{
A \code{targets} pipeline output, typically a nested tibble with differential abundance estimates.
}
\description{
This function prepares and runs a differential abundance test pipeline using the 'targets' package. It sets up necessary files, appends scripts, and executes the pipeline.
}
