% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map_test_differential_abundance}
\alias{map_test_differential_abundance}
\title{Test Differential Abundance in SummarizedExperiment Object}
\usage{
map_test_differential_abundance(
  se,
  .col,
  .formula,
  .abundance = NULL,
  max_rows_for_matrix_multiplication = NULL,
  cores = 1,
  ...
)
}
\arguments{
\item{se}{Data frame containing SummarizedExperiment objects.}

\item{.col}{Column in the data frame containing the SummarizedExperiment objects.}

\item{.formula}{Formula for the differential abundance test.}

\item{.abundance}{Optional; a symbol indicating the column of the datasets that specifies the abundance measure. If NULL, a default is used.}

\item{max_rows_for_matrix_multiplication}{Maximum number of rows for matrix multiplication.}

\item{cores}{Number of cores to use for computation.}

\item{...}{Additional parameters.}
}
\value{
Data frame with test results.
}
\description{
Applies differential abundance testing to each SummarizedExperiment object in a data frame.
}
