% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{non_batch_variation_removal}
\alias{non_batch_variation_removal}
\title{Non_batch_variation_removal}
\usage{
non_batch_variation_removal(
  input_path_demultiplexed,
  input_path_empty_droplets,
  alive_identification_tbl,
  cell_cycle_score_tbl
)
}
\description{
Regress out variations due to mitochondrial content, ribosomal content, and 
cell cycle effects
}
