% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{non_batch_variation_removal}
\alias{non_batch_variation_removal}
\title{Non-Batch Variation Removal}
\usage{
non_batch_variation_removal(
  input_read_RNA_assay,
  empty_droplets_tbl = NULL,
  alive_identification_tbl = NULL,
  doublet_identification_tbl = NULL,
  cell_cycle_score_tbl = NULL,
  assay = NULL,
  factors_to_regress = NULL,
  external_path,
  container_type
)
}
\arguments{
\item{input_read_RNA_assay}{A \code{SingleCellExperiment} or \code{Seurat} object containing RNA assay data.}

\item{empty_droplets_tbl}{A tibble identifying empty droplets.}

\item{alive_identification_tbl}{A tibble from alive cell identification.}

\item{doublet_identification_tbl}{A tibble from doublet identification.}

\item{cell_cycle_score_tbl}{A tibble from cell cycle scoring.}

\item{assay}{assay used, default = "RNA"}

\item{factors_to_regress}{A character vector specifying cell-level covariates to regress out during \code{SCTransform} normalization.}

\item{external_path}{A character vector of directory that the data to be stored}

\item{data_container_type}{A character vector specifying the output file type. Ideally it should match to the input file type.}
}
\value{
Normalized and adjusted data.
}
\description{
Regresses out variations due to mitochondrial content, ribosomal content, and
cell cycle effects.
}
