% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{preprocessing_output}
\alias{preprocessing_output}
\title{Preprocessing_output}
\usage{
preprocessing_output(
  tissue,
  non_batch_variation_removal_S,
  alive_identification_tbl,
  cell_cycle_score_tbl,
  annotation_label_transfer_tbl,
  doublet_identification_tbl
)
}
\description{
Incorporates outputs from doublets and dead cell filtering, cell cycle scoring, 
and optionally includes annotation label transfer information. Generate a 
processed dataset ready for downstream analysis.
}
