% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pseudobulk_merge}
\alias{pseudobulk_merge}
\title{Merge pseudobulk from all samples}
\usage{
pseudobulk_merge(create_pseudobulk_sample, assays, x, ...)
}
\arguments{
\item{create_pseudobulk_sample}{A list pseudobulk samples generated by \code{create_pseudobulk}}

\item{assays}{Default is set of \code{RNA}}

\item{x}{User specified character vector for the column from which we subset the samples for pseudobulk analysis}

\item{...}{Additional arguments}
}
\description{
Merge pseudobulk from all samples. Ensures that missing genes are accounted
for and aligns data across multiple samples.
}
