% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pseudobulk_preprocessing}
\alias{pseudobulk_preprocessing}
\title{Pseudobulk Preprocessing}
\usage{
pseudobulk_preprocessing(
  reference_label_fine,
  preprocessing_output_S,
  sample_column
)
}
\arguments{
\item{reference_label_fine}{Reference label for fine categorization.}

\item{preprocessing_output_S}{Processed dataset from preprocessing.}

\item{sample_column}{Column name indicating sample identifiers.}
}
\value{
List containing pseudobulk data aggregated by sample and by both sample and cell type.
}
\description{
Aggregates cells based on sample and cell type annotations, creating pseudobulk samples
for each combination. Handles RNA and ADT assays, ensuring that missing genes are accounted
for and aligns data across multiple samples.
}
