% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_data_container}
\alias{read_data_container}
\title{Read various types of single-cell data}
\usage{
read_data_container(file, container_type = "anndata")
}
\arguments{
\item{file}{A vector of file path}

\item{container_type}{A character vector of length one specifies the input data type.}
}
\value{
A \code{Seurat} object
}
\description{
Read various types of single-cell data
}
