% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reference_label_fine_id}
\alias{reference_label_fine_id}
\title{Reference Label Fine Identification}
\usage{
reference_label_fine_id(tissue)
}
\arguments{
\item{tissue}{Type of tissue.}
}
\value{
Appropriate reference label for fine categorization.
}
\description{
Adds a reference label depending on the provided tissue type of the input data.
}
