% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_pipeline.R
\name{run_targets_pipeline}
\alias{run_targets_pipeline}
\title{Run Targets Pipeline for HPCell}
\usage{
run_targets_pipeline(
  input_data,
  store = "./",
  input_reference = NULL,
  tissue,
  computing_resources = crew_controller_local(workers = 1),
  debug_step = NULL,
  filter_empty_droplets = NULL,
  RNA_assay_name = "RNA",
  sample_column = "sample",
  cell_type_annotation_column = "Cell_type_in_each_tissue"
)
}
\arguments{
\item{input_data}{Input data for the pipeline.}

\item{store}{Directory path for storing the pipeline files.}

\item{input_reference}{Optional reference data.}

\item{tissue}{Tissue type for the analysis.}

\item{computing_resources}{Configuration for computing resources.}

\item{debug_step}{Optional step for debugging.}

\item{filter_empty_droplets}{Flag to indicate if input filtering is needed.}

\item{RNA_assay_name}{Name of the RNA assay.}

\item{sample_column}{Column name for sample identification.}

\item{cell_type_annotation_column}{Column name for cell type annotation in input data}
}
\value{
The output of the \code{targets} pipeline, typically a pre-processed data set.
}
\description{
This function sets up and executes a \code{targets} pipeline for HPCell. It saves input data and configurations,
writes a pipeline script, and runs the pipeline using the 'targets' package.
}
