% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{split_sample_cell_type_calculate_metacell_membership}
\alias{split_sample_cell_type_calculate_metacell_membership}
\title{Calculate Metacell Membership for Each Cell Type}
\usage{
split_sample_cell_type_calculate_metacell_membership(sample_sce, cell_type)
}
\arguments{
\item{sample_sce}{A SingleCellExperiment object containing single-cell data.}

\item{cell_type}{The variable from the colData of \code{sample_sce} used to group cells by type.}
}
\value{
A tibble with metacell membership data for each cell type.
}
\description{
This function processes a SingleCellExperiment object by grouping cells according
to their type, calculates metacell membership for each group, and combines the
results into a single tibble.
}
