% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_grammar_hpc.R
\name{test_differential_abundance-HPCell-method}
\alias{test_differential_abundance-HPCell-method}
\title{Test Differential Abundance for HPCell}
\usage{
\S4method{test_differential_abundance}{HPCell}(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  contrasts = NULL,
  method = "edgeR_quasi_likelihood",
  test_above_log2_fold_change = NULL,
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  prefix = "",
  action = "add",
  factor_of_interest = NULL,
  target_input = "pseudobulk_se",
  target_output = "de",
  group_by_column = NULL,
  ...,
  significance_threshold = NULL,
  fill_missing_values = NULL,
  .contrasts = NULL
)
}
\arguments{
\item{.data}{An HPCell object.}

\item{.formula}{A formula used to model the design matrix.}

\item{.sample}{Sample parameter.}

\item{.transcript}{Transcript parameter.}

\item{.abundance}{Abundance parameter.}

\item{contrasts}{Contrasts parameter.}

\item{method}{Method parameter, default is "edgeR_quasi_likelihood".}

\item{test_above_log2_fold_change}{Test above log2 fold change.}

\item{scaling_method}{Scaling method, default is "TMM".}

\item{omit_contrast_in_colnames}{Omit contrast in column names.}

\item{prefix}{Prefix parameter.}

\item{action}{Action parameter, default is "add".}

\item{...}{Additional parameters.}

\item{significance_threshold}{Significance threshold.}

\item{fill_missing_values}{Fill missing values.}

\item{.contrasts}{Contrasts parameter.}
}
\value{
The result of the differential abundance test.
}
\description{
This function tests differential abundance for HPCell objects.
}
