% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tranform_assay.R
\name{transform_utility}
\alias{transform_utility}
\title{Apply a transformation to an assay and save as HDF5}
\usage{
transform_utility(input_read_RNA_assay, transform_fx, external_path)
}
\arguments{
\item{input_read_RNA_assay}{A SummarizedExperiment object to be transformed.}

\item{external_path}{A character string specifying the directory path to save the transformed object.}

\item{transform}{A function to apply to the assay of the SummarizedExperiment object.}
}
\value{
The function does not return an object. It saves the transformed SummarizedExperiment object to the specified path.
}
\description{
This function applies a specified transformation to the assay of a
SummarizedExperiment object and saves the transformed object in HDF5 format.
}
